moved redirect functions to fileIO.py
