+++ baseclasses/utils/__init__.py
@@ -1,8 +1,7 @@
 from .containers import CaseInsensitiveSet, CaseInsensitiveDict
 from .error import Error
 from .utils import getPy3SafeString, pp
-from .fileIO import writeJSON, readJSON, writePickle, readPickle
-from .redirectIO import redirectIO, redirectingIO
+from .fileIO import writeJSON, readJSON, writePickle, readPickle, redirectIO, redirectingIO
 
 __all__ = [
     "CaseInsensitiveSet",
